# Improve speed of transformPointCloud/WithNormals() functions