# tl;dr;

This adds an implementation of point cloud transform functions using SSE2 intrinsics. Depending on the transform scalar precision, compiler flags, compiler version, point type, and point cloud properties this may lead to up to 35% faster transforms of VGA-sized point clouds.

The old, non-optimized implementation is retained as a fallback for machines without SSE2 instructions.

# Long version

Recently I was wondering if I can improve performance of `transformPointCloud()` family of functions. This is how they transform every point internally:

https://github.com/PointCloudLibrary/pcl/blob/c78482bacb515c3880386c6ab4566ca964c2ebe2/common/include/pcl/common/impl/transforms.hpp#L67-L71

This manually spelled out vector-matrix product looks strange, especially since Eigen should properly align points and provide vectorized matrix operations. In fact, this already seemed strange to me nearly four years ago. So I asked why and the [answer](https://github.com/PointCloudLibrary/pcl/commit/0266f79136d8c214628ba8d86a02bb2a08a431d8#r7772811) was:

> You wouldn't believe it, but we had a look a long time ago at the generated assembly code and it was faster. Things might have change since then.

I decided to check myself. My development environment is Ubuntu 16.04 with Eigen 3.3 and GCC 5.4. When I switched to use Eigen's vector-matrix multiplication operator, I observed 2x slowdown, indeed. Same with GCC 6.3. However, GCC 7.2 gave nearly 2x speeedup! Clearly, older compilers were not vectorizing properly.

Looking at the disassembly, I found that indeed they are not using vectorized SSE2 instructions. Funny enough, when I compiled with GCC 5.4 and `-msse2` flag (instead of `-march=native`), I got vectorized code. Turns out, in native mode in addition to SSE2 extensions, FMA (fuse-multiply-add) instructions become enabled and compiler starts to abuse them.

So simply switching to vector-matrix product is not an option because depending on the compiler version and compiler flags this introduces either speedups or slowdowns. An alternative was to implement everything directly with SSE2 intrinsics such that compiler can not screw up and is guaranteed to emit optimal SSE2 assembly.

Since PCL does not have a built-in benchmarking framework, I have a separate [repository](https://github.com/taketwo/pcl-transforms) with the proposed implementation and benchmarks. Here are my results with VGA-sized point clouds on i7 from 2015:
 
![benchmark](https://user-images.githubusercontent.com/1241736/37096344-a292a0f6-2218-11e8-9d74-31b5104213e4.png)

For every cell two tests are run: baseline PCL transform and proposed transform. The reported number is the runtime (in microseconds) for the proposed transform and the fraction of the baseline time. Thus numbers less than 1.0 represent improved performance.

I am lazy to provide complete discussion of the results and let anyone interested study this table or even run tests himself. Just couple of points:

 1. The "-no-sse2" people are obviously not affected by this change;
 2. Dense `XYZ` point clouds are transformed 24% to 35% faster;
 3. For double-precision transform matrices there is a performance gain only with old GCC. There are a couple of red cells also, but I tend to think these are measurement noise;
 4. Transformation of `XYZRGBNormal` point clouds is nearly 3 times slower than `XYZ` clouds, although it involves only double the number of operations. I conclude that memory access is the real bottleneck here.

Closes #1255.

<strike>One last note: double-precision transforms can be made faster with AVX instructions. But I leave this as an exercise for the future generations.</strike> (This PR now contains AVX optimizations as well.)