@SergioRAgostinho 2018-03-07
I'll have three machines in which I'm interested in giving this benchmark a try. I'll report results later.