@SergioRAgostinho 2018-03-07
**CPU**: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
**OS**: Ubuntu 16.04
**Kernel**: 4.13.0-36-generic
**Compiler**: gcc-5-4-0

### None
```
Running benchmarks
Compiler: gcc-5-4-0
Flag: none
Celero
Timer resolution: 0.001000 us
Writing results to: out
-----------------------------------------------------------------------------------------------------------------------------------------------
     Group      |   Experiment    |   Prob. Space   |     Samples     |   Iterations    |    Baseline     |  us/Iteration   | Iterations/sec  | 
-----------------------------------------------------------------------------------------------------------------------------------------------
D_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1143.55000 |          874.47 | 
D_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2458.16000 |          406.81 | 
D_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1455.81000 |          686.90 | 
D_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3523.26000 |          283.83 | 
S_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1217.54000 |          821.33 | 
S_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2541.50000 |          393.47 | 
S_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1487.62000 |          672.21 | 
S_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3589.68000 |          278.58 | 
D_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.91872 |      1050.60000 |          951.84 | 
D_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.79790 |      1961.37000 |          509.85 | 
D_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.91492 |      1331.95000 |          750.78 | 
D_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.85674 |      3018.53000 |          331.29 | 
S_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.80564 |       980.90000 |         1019.47 | 
S_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.70550 |      1793.02000 |          557.72 | 
S_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.84202 |      1252.61000 |          798.33 | 
S_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.78879 |      2831.50000 |          353.17 | 
Complete.
```

### Native
```
Flag: native
Celero
Timer resolution: 0.001000 us
Writing results to: out
-----------------------------------------------------------------------------------------------------------------------------------------------
     Group      |   Experiment    |   Prob. Space   |     Samples     |   Iterations    |    Baseline     |  us/Iteration   | Iterations/sec  | 
-----------------------------------------------------------------------------------------------------------------------------------------------
D_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |       888.64000 |         1125.32 | 
D_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2033.82000 |          491.69 | 
D_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1074.31000 |          930.83 | 
D_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      2984.99000 |          335.01 | 
S_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1046.48000 |          955.58 | 
S_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2276.48000 |          439.27 | 
S_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1304.99000 |          766.29 | 
S_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3348.83000 |          298.61 | 
D_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.64262 |       571.06000 |         1751.13 | 
D_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.70225 |      1428.24000 |          700.16 | 
D_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.78740 |       845.91000 |         1182.16 | 
D_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.81589 |      2435.42000 |          410.61 | 
S_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.75009 |       784.95000 |         1273.97 | 
S_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.63527 |      1446.17000 |          691.48 | 
S_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.79945 |      1043.27000 |          958.52 | 
S_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.74284 |      2487.64000 |          401.99 | 
Complete.
```

### SSE2

```
Flag: sse2
Celero
Timer resolution: 0.001000 us
Writing results to: out
-----------------------------------------------------------------------------------------------------------------------------------------------
     Group      |   Experiment    |   Prob. Space   |     Samples     |   Iterations    |    Baseline     |  us/Iteration   | Iterations/sec  | 
-----------------------------------------------------------------------------------------------------------------------------------------------
D_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1134.42000 |          881.51 | 
D_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2434.32000 |          410.79 | 
D_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1455.29000 |          687.15 | 
D_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3526.78000 |          283.54 | 
S_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1208.06000 |          827.77 | 
S_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2525.06000 |          396.03 | 
S_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1492.27000 |          670.12 | 
S_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3570.31000 |          280.09 | 
D_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.92557 |      1049.98000 |          952.40 | 
D_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.80552 |      1960.89000 |          509.97 | 
D_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.91754 |      1335.29000 |          748.90 | 
D_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.86764 |      3059.96000 |          326.80 | 
S_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.82094 |       991.75000 |         1008.32 | 
S_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         0.70866 |      1789.40000 |          558.85 | 
S_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.84115 |      1255.23000 |          796.67 | 
S_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         0.79985 |      2855.72000 |          350.17 | 
Complete.
```

### NO-SSE2

```
Flag: no-sse2
Celero
Timer resolution: 0.001000 us
Writing results to: out
-----------------------------------------------------------------------------------------------------------------------------------------------
     Group      |   Experiment    |   Prob. Space   |     Samples     |   Iterations    |    Baseline     |  us/Iteration   | Iterations/sec  | 
-----------------------------------------------------------------------------------------------------------------------------------------------
D_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |       813.59000 |         1229.12 | 
D_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      1799.12000 |          555.83 | 
D_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1110.37000 |          900.60 | 
D_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      2892.70000 |          345.70 | 
S_XYZ_NC        | PCL             |            Null |              50 |             100 |         1.00000 |      1048.35000 |          953.88 | 
S_XYZRGBN_NC    | PCL             |            Null |              50 |             100 |         1.00000 |      2066.47000 |          483.92 | 
S_XYZ_C         | PCL             |            Null |              50 |             100 |         1.00000 |      1322.48000 |          756.16 | 
S_XYZRGBN_C     | PCL             |            Null |              50 |             100 |         1.00000 |      3124.52000 |          320.05 | 
D_XYZ_NC        | Proposed        |            Null |              50 |             100 |         0.99837 |       812.26000 |         1231.13 | 
D_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         1.08292 |      1948.30000 |          513.27 | 
D_XYZ_C         | Proposed        |            Null |              50 |             100 |         0.98999 |      1099.26000 |          909.70 | 
D_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         1.04916 |      3034.91000 |          329.50 | 
S_XYZ_NC        | Proposed        |            Null |              50 |             100 |         1.03451 |      1084.53000 |          922.06 | 
S_XYZRGBN_NC    | Proposed        |            Null |              50 |             100 |         1.10091 |      2274.99000 |          439.56 | 
S_XYZ_C         | Proposed        |            Null |              50 |             100 |         1.01860 |      1347.08000 |          742.35 | 
S_XYZRGBN_C     | Proposed        |            Null |              50 |             100 |         1.06407 |      3324.70000 |          300.78 | 
Complete.
```